requirements.txt
openai
weave
markdown
bleach
python-fasthtml
instructor